move rust kernels to module
