+++ daft/internal/kernels/hashing.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pyarrow as pa
 
-from daft import daft as _daft
+from daft.daft import kernels
 
 
 def hash_chunked_array(arr, seed=None):
@@ -15,10 +15,10 @@ def hash_chunked_array(arr, seed=None):
         assert len(seed) == len(arr)
         for chunk in arr.chunks:
             result_chunks.append(
-                _daft.hash_pyarrow_array(chunk, pa, seed=seed.slice(offset, len(chunk)).combine_chunks())
+                kernels.hash_pyarrow_array(chunk, pa, seed=seed.slice(offset, len(chunk)).combine_chunks())
             )
             offset += len(chunk)
     else:
         for chunk in arr.chunks:
-            result_chunks.append(_daft.hash_pyarrow_array(chunk, pa, seed=None))
+            result_chunks.append(kernels.hash_pyarrow_array(chunk, pa, seed=None))
     return pa.chunked_array(result_chunks, type=pa.uint64())
